refactor(repair): consolidate merge-readiness GitHub reads
