+++ src/repair/merge-readiness-github.ts
@@ -1,0 +1,79 @@
+import type { JsonValue } from "./json-types.js";
+import { ghJsonWithRetry as ghJson } from "./github-cli.js";
+
+export function validateResolvedReviewThreads(repo: string, number: JsonValue) {
+  const [owner, name] = repo.split("/");
+  const query = `
+    query($owner: String!, $name: String!, $number: Int!) {
+      repository(owner: $owner, name: $name) {
+        pullRequest(number: $number) {
+          reviewThreads(first: 100) {
+            pageInfo { hasNextPage }
+            nodes {
+              isResolved
+              path
+              line
+              comments(first: 1) {
+                nodes {
+                  url
+                  author { login }
+                  body
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  `;
+  const data = ghJson([
+    "api",
+    "graphql",
+    "-f",
+    `owner=${owner}`,
+    "-f",
+    `name=${name}`,
+    "-F",
+    `number=${number}`,
+    "-f",
+    `query=${query}`,
+  ]);
+  const threads = data?.data?.repository?.pullRequest?.reviewThreads;
+  if (threads?.pageInfo?.hasNextPage) return "too many review threads to prove resolved";
+  const unresolved = (threads?.nodes ?? []).filter(
+    (thread: JsonValue) => thread && !thread.isResolved,
+  );
+  if (unresolved.length === 0) return "";
+  const examples = unresolved
+    .slice(0, 3)
+    .map(
+      (thread: JsonValue) =>
+        thread.comments?.nodes?.[0]?.url ?? `${thread.path}:${thread.line ?? "?"}`,
+    );
+  return `unresolved review threads remain: ${examples.join(", ")}`;
+}
+
+export function fetchPullRequestView(repo: string, number: JsonValue) {
+  return ghJson([
+    "pr",
+    "view",
+    String(number),
+    "--repo",
+    repo,
+    "--json",
+    [
+      "baseRefName",
+      "isDraft",
+      "mergeable",
+      "mergeCommit",
+      "mergeStateStatus",
+      "mergedAt",
+      "reviewDecision",
+      "state",
+      "statusCheckRollup",
+      "title",
+      "updatedAt",
+      "url",
+    ].join(","),
+  ]);
+}
